+++ swe400fitness/src/commands/CommandToRejectFriendRequest.java
@@ -23,7 +23,7 @@ public class CommandToRejectFriendRequest implements Command
 	 * @param userIDOfRequestee the User ID of the user accepting the request
 	 * @param userNameOfRequester the User Name of the user who initiated the friend request
 	 */
-	public CommandToRejectFriendRequest(int userIDOfRequestee, String userNameOfRequester)
+	public CommandToRejectFriendRequest(long userIDOfRequestee, String userNameOfRequester)
 	{
 		this.userIDOfRequestee = userIDOfRequestee;
 		this.userNameOfRequester = userNameOfRequester;
